Reduced repetition by adding mixin class
