+++ osler/assertion.py
@@ -1,12 +1,9 @@
 #!/usr/bin/env python3
 
-import logging
-log = logging.getLogger(__name__)
-
-from .graph import Node
+from .common import EntityBase, NodeMixin
 from .criterion import Criterion
 
-class Assertion(object):
+class Assertion(EntityBase, NodeMixin):
 
     def __init__(self, proposition, test_difficulty=0.0, **kwargs):
         self.proposition = proposition
@@ -17,15 +14,6 @@ class Assertion(object):
             self.cannot_preceed = kwargs.get('cannot_preceed')
         else:
             self.cannot_preceed = None
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return self.__str__()
-
-    def __hash__(self):
-        return hash(str(self))
 
     def __eq__(self, other):
         if isinstance(other, Assertion):
@@ -38,6 +26,3 @@ class Assertion(object):
     
     def false(self):
         return Criterion(self, False)
-
-    def parent(self, parent_node):
-        return Node(self, parent_node)
